# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Alyra Solutions - Multi-tenant Nail Salon SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salons:
    get:
      operationId: listSalons
      tags: [salons]
      summary: List all salons (platform admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        "200":
          description: List of salons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonListResponse"

    post:
      operationId: createSalon
      tags: [salons]
      summary: Create a new salon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalonRequest"
      responses:
        "201":
          description: Salon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
  /salons/{salonId}:
    get:
      operationId: getSalon
      tags: [salons]
      summary: Get salon details
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: Salon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"

    put:
      operationId: updateSalon
      tags: [salons]
      summary: Update salon
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalonRequest"
      responses:
        "200":
          description: Salon updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"

    delete:
      operationId: deleteSalon
      tags: [salons]
      summary: Delete salon
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: Salon deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /salons/by-subdomain/{subdomain}:
    get:
      operationId: getSalonBySubdomain
      tags: [salons]
      summary: Get salon by subdomain
      parameters:
        - in: path
          name: subdomain
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Salon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
  /salons/{salonId}/stats:
    get:
      operationId: getSalonStats
      tags: [salons]
      summary: Get salon statistics
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: Salon stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonStats"

  # Users
...
  /salons/{salonId}/services:
    get:
      operationId: listServices
      tags: [services]
      summary: List salon services
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"

    post:
      operationId: createService
      tags: [services]
      summary: Create service
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /salons/{salonId}/services/{serviceId}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service details
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/serviceId"
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    put:
      operationId: updateService
      tags: [services]
      summary: Update service
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/serviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/serviceId"
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Appointments
...
  /admin/salons:
    get:
      operationId: adminListSalons
      tags: [salons]
      summary: Admin - list all salons
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All salons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalonListResponse"
...
components:
  parameters:
    salonId:
      in: path
      name: salonId
      required: true
      schema:
        type: integer
...
    serviceId:
      in: path
      name: serviceId
      required: true
      schema:
        type: integer
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # Auth
...
    Salon:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subdomain:
          type: string
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zipCode:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended]
        timezone:
          type: string
        openingHours:
          $ref: "#/components/schemas/OpeningHours"
        ownerId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - subdomain
        - status
        - timezone
        - ownerId
        - createdAt
        - updatedAt
    OpeningHours:
      type: object
      properties:
        monday:
          $ref: "#/components/schemas/DayHours"
        tuesday:
          $ref: "#/components/schemas/DayHours"
        wednesday:
          $ref: "#/components/schemas/DayHours"
        thursday:
          $ref: "#/components/schemas/DayHours"
        friday:
          $ref: "#/components/schemas/DayHours"
        saturday:
          $ref: "#/components/schemas/DayHours"
        sunday:
          $ref: "#/components/schemas/DayHours"
    DayHours:
      type: object
      properties:
        open:
          type: boolean
        openTime:
          type: string
          nullable: true
        closeTime:
          type: string
          nullable: true
      required:
        - open
    CreateSalonRequest:
      type: object
      properties:
        name:
          type: string
        subdomain:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        timezone:
          type: string
        ownerId:
          type: integer
      required:
        - name
        - subdomain
        - timezone
        - ownerId
    UpdateSalonRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        timezone:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        openingHours:
          $ref: "#/components/schemas/OpeningHours"
    SalonListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Salon"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    SalonStats:
      type: object
      properties:
        totalAppointments:
          type: integer
        todayAppointments:
          type: integer
        pendingAppointments:
          type: integer
        totalClients:
          type: integer
        totalTechnicians:
          type: integer
        currentWaitlistCount:
          type: integer
        avgWaitTime:
          type: integer
        monthlyRevenue:
          type: number
        totalRevenue:
          type: number
      required:
        - totalAppointments
        - todayAppointments
        - pendingAppointments
        - totalClients
        - totalTechnicians
        - currentWaitlistCount
        - avgWaitTime
        - monthlyRevenue
        - totalRevenue

    # Technicians
...
    Service:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        price:
          type: number
        isActive:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - name
        - category
        - duration
        - price
        - isActive
        - createdAt
        - updatedAt
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        duration:
          type: integer
        price:
          type: number
      required:
        - name
        - category
        - duration
        - price
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        duration:
          type: integer
        price:
          type: number
        isActive:
          type: boolean
    ServiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # Appointments
...