# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alyra Solutions - Multi-tenant Nail Salon SaaS Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface Service {
  id: number;
  salonId: number;
  name: string;
  description?: string | null;
  category: string;
  /** Duration in minutes */
  duration: number;
  price: number;
  isActive: boolean;
  imageUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}